changeLog: (tested on A7s) *** A full list of camera ISOs is required in cameraModels.json *** See the A7s for an example.
imageReady now returns an array of int instead of ushort.
Updated imageProcessor to use int instead of uint throughout, not sure if this will cause issues as there are no tests.

No longer assumes bulb mode is available.
If CanStopExposure is set to false EndExposure will now wait for exposure to finish.
StopExposure will throw if called and CanStopExposure is set to false

In cameraModels.json
AllGain is a list of all available ISOs on camera
Gain is a list available to the program using the ASCOM driver (limited to 25600 as its stored as a short).

Shutter speed and ISO are now pressed the specific number of times needed to get to wanted value, then verified as UI takes a while to update to the current value set on the camera.